requirements.txt
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-bcrypt==0.7.1
python-dotenv==0.19.2
email-validator==1.1.3
gunicorn==20.1.0
pytest==6.2.5
